requirements.txt
Django==2.1.4
django-cors-middleware==1.3.1
django-extensions==2.1.4
djangorestframework==3.9.0
PyJWT==1.4.2
six==1.10.0
